Use existing code to close per 5 minutes gaps.

+++ app/src/main/java/com/eveningoutpost/dexdrip/Models/OOPResults.java
@@ -10,7 +10,7 @@ class HistoricBg {
 
     public int quality;
     public int time;
-    public int bg;
+    public double bg;
 }
 
 public class OOPResults {
